polishing up

+++ MatchEngine/Player.m
@@ -32,7 +32,7 @@
 @synthesize nameTextField;
 
 
-- (id)init:(NSInteger)p_id number:(NSInteger)num jersey:(NSInteger)jer team:(NSInteger)t role:(NSInteger)r name:(NSString*)n
+- (id)init:(NSInteger)p_id number:(NSInteger)num jersey:(NSInteger)jer team:(int)t role:(NSInteger)r name:(NSString*)n
 {
     if (self = [super init]) 
     {
@@ -76,7 +76,7 @@
     return self;
 }
 
-+ (Player *)initPlayer:(NSInteger)player_id number:(NSInteger)number jersey:(NSInteger)jersey team:(NSInteger)team role:(NSInteger)role name:(NSString*)name
++ (Player *)initPlayer:(NSInteger)player_id number:(NSInteger)number jersey:(NSInteger)jersey team:(int)team role:(NSInteger)role name:(NSString*)name
 {
 	return [[Player alloc] init:player_id number:number jersey:jersey team:team role:role name:name];
 }
